+++ app/scripts/lib/config-manager.js
@@ -102,6 +102,17 @@ ConfigManager.prototype.setWallet = function(wallet) {
   this.setData(data)
 }
 
+ConfigManager.prototype.getSelectedAccount = function() {
+  var config = this.getConfig()
+  return config.selectedAccount
+}
+
+ConfigManager.prototype.setSelectedAccount = function(address) {
+  var config = this.getConfig()
+  config.selectedAccount = address
+  this.setConfig(config)
+}
+
 ConfigManager.prototype.getWallet = function() {
   return this.migrator.getData().wallet
 }