+++ ui/app/info.js
@@ -26,7 +26,7 @@ InfoScreen.prototype.render = function() {
       h('.section-title.flex-row.flex-center', [
         h('i.fa.fa-arrow-left.fa-lg.cursor-pointer', {
           onClick: (event) => {
-            state.dispatch(actions.showAccountsPage())
+            state.dispatch(actions.goHome())
           }
         }),
         h('h2.page-subtitle', 'Info'),